the number filter returns!
alright, so deleting the reversed elements broke the thing. we've returned the numpair filter, no idea if it has any effect on performance tho. in other news, we've also removed the "if not skolem_arr" that comes after the array is generated. as this value should never return true for any value of k that's valid, i've removed it.